+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+## Added
+- Display keybindings in scrubber tooltips
 
 ## [0.3.1] - 2016-08-01
 ### Fixed