+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.4.0] - 2016-08-07
 ## Added
 - Display keybindings in scrubber tooltips
 - `Launcher.cardArt()` and `.assets()` can now accept a callback
@@ -53,7 +55,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Code for initial development release
 
-[Unreleased]: https://github.com/HearthSim/joust/compare/0.3.1...HEAD
+[Unreleased]: https://github.com/HearthSim/joust/compare/0.4.0...HEAD
 [0.2.0]: https://github.com/HearthSim/joust/compare/0.1.0...0.2.0
 [0.3.0]: https://github.com/HearthSim/joust/compare/0.2.0...0.3.0
 [0.3.1]: https://github.com/HearthSim/joust/compare/0.3.0...0.3.1
+[0.4.0]: https://github.com/HearthSim/joust/compare/0.3.1...0.4.0